Run the spec suite over several worker processes
Nearly every example boots a Rails, Hanami or Roda app in a child process
and compares the document it generates against a committed fixture, so the
suite is a queue of child processes rather than CPU-bound Ruby. Run serially
it leaves every core but one idle, and the Ruby 4.0 job kept brushing the CI
step timeout because it is the one job that also collects coverage, which
costs another 25% per spawned process.

scripts/parallel_rspec spreads the spec files over one worker per core. No
two spec files write the same document, so this is safe. Two changes make it
so in practice.

spec/rspec/rails_spec.rb held eleven groups and took as long as the rest of
the suite together, which capped what parallelism could buy. Its groups own
separate fixtures, so the ones covering per-version output and merging into
an existing document move to their own files. Only the three groups that
share rspec_openapi.yaml and rspec_openapi.json stay together.

The upload examples wrote test.png into the repository root from three
different suites, which concurrent runs would read half-written. They now
share spec/support/upload_fixture.rb, which gives each process its own
directory. The documents still record the file as test.png.

Also pass -r./.simplecov_spawn to the child spawned from rails_test.rb. It
was the one spawn without it, so it fell back to the default SimpleCov
formatter and wrote a full HTML report over the whole merged result set.

Coverage is unchanged: line-rate 0.9872, branch-rate 0.8691 over 27 files,
the same as a serial run.
