+++ .gitignore
@@ -17,5 +17,3 @@ spec/apps/rails/log/
 
 /mise.toml
 
-# Test artifact generated by request specs (e.g. spec/requests/rails_spec.rb)
-/test.png
