+++ spec/rspec/rails_merge_spec.rb
@@ -1,0 +1,80 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+require 'yaml'
+require 'json'
+require 'pry'
+
+RSpec.describe 'rails request spec, merging into an existing document' do
+  include SpecHelper
+
+  describe 'hand-edited 3.2 document round-trip' do
+    let(:input_path) do
+      File.expand_path('spec/apps/rails/doc/roundtrip/input.yaml', repo_root)
+    end
+
+    let(:expected_path) do
+      File.expand_path('spec/apps/rails/doc/roundtrip/expected.yaml', repo_root)
+    end
+
+    # The seed document carries JSON-Schema null type arrays on an untouched
+    # path. They are normalized while reading (then dropped from the output, as
+    # any un-recorded path is), so the regenerated file holds only the recorded
+    # path. We restore the seed afterwards.
+    it 'normalizes null type arrays read from an existing document' do
+      original_source = File.read(input_path)
+      begin
+        rspec 'spec/requests/rails_3_2_roundtrip_spec.rb', openapi: true, output: :yaml
+        new_yaml = YAML.safe_load(File.read(input_path))
+        expected_yaml = YAML.safe_load(File.read(expected_path))
+        expect(new_yaml).to eq expected_yaml
+      ensure
+        File.write(input_path, original_source)
+      end
+    end
+  end
+
+  describe 'smart merge' do
+    let(:openapi_path) do
+      File.expand_path('spec/apps/rails/doc/smart/openapi.yaml', repo_root)
+    end
+
+    let(:expected_path) do
+      File.expand_path('spec/apps/rails/doc/smart/expected.yaml', repo_root)
+    end
+
+    it 'updates the spec/apps/rails/doc/smart/openapi.yaml as same as in expected.yaml' do
+      original_source = File.read(openapi_path)
+      begin
+        rspec 'spec/requests/rails_smart_merge_spec.rb', openapi: true, output: :yaml
+        new_yaml = YAML.safe_load(File.read(openapi_path))
+        expected_yaml = YAML.safe_load(File.read(expected_path))
+        expect(new_yaml).to eq expected_yaml
+      ensure
+        File.write(openapi_path, original_source)
+      end
+    end
+  end
+
+  describe 'description preservation with example_mode :none' do
+    let(:openapi_path) do
+      File.expand_path('spec/apps/rails/doc/description_preserve/openapi.yaml', repo_root)
+    end
+
+    let(:expected_path) do
+      File.expand_path('spec/apps/rails/doc/description_preserve/expected.yaml', repo_root)
+    end
+
+    it 'preserves existing descriptions when example_mode is :none but overwrites with normal mode' do
+      original_source = File.read(openapi_path)
+      begin
+        rspec 'spec/requests/rails_description_preserve_spec.rb', openapi: true, output: :yaml
+        new_yaml = YAML.safe_load(File.read(openapi_path))
+        expected_yaml = YAML.safe_load(File.read(expected_path))
+        expect(new_yaml).to eq expected_yaml
+      ensure
+        File.write(openapi_path, original_source)
+      end
+    end
+  end
+end
